+++ particle.h
@@ -19,39 +19,39 @@ typedef enum {
 class Particle {
 	public:
 		ThreeDVector* position;
-		long double mass;
+		float mass;
 		ThreeDVector* velocity;
 		ThreeDVector* velocity_half;
-		long double viscosity_coefficient;
-		long double gas_constant;
-		long double buoyancy_strength;
-		long double rest_density;
-		long double temperature;
+		float viscosity_coefficient;
+		float gas_constant;
+		float buoyancy_strength;
+		float rest_density;
+		float temperature;
 
 		 Particle_Type type;
 
-		static map<ThreeDVector*, long double, comparator>* color_map;
+		static map<ThreeDVector*, float, comparator>* color_map;
 
 		//Interpolated Fields
-		long double density;
+		float density;
 		ThreeDVector* acceleration;
 
-		Particle(long double x, long double y, long double z, long double mass, 
-			ThreeDVector* velocity, long double viscosity_coefficient, long double buoyancy_strength, 
-			long double gas_constant, long double rest_density, long double temperature, Particle_Type t);
+		Particle(float x, float y, float z, float mass, 
+			ThreeDVector* velocity, float viscosity_coefficient, float buoyancy_strength, 
+			float gas_constant, float rest_density, float temperature, Particle_Type t);
 		~Particle();
 		void set_density(vector<Particle*>* particles);
 		void set_acceleration(vector<Particle*>* particles);
 
 		//Leapfrog steps from http://www.cs.cornell.edu/~bindel/class/cs5220-f11/code/sph.pdf
-		void leapfrog_start(long double dt);
-		void leapfrog_step(long double dt);
+		void leapfrog_start(float dt);
+		void leapfrog_step(float dt);
 		
 		
 		ThreeDVector* viscosityForce(vector<Particle*>* particles);
 		ThreeDVector* pressureForce(vector<Particle*>* particles);
 		ThreeDVector* boundaryForce(vector<Particle*>* particles);
-		long double pressure();
+		float pressure();
 		ThreeDVector* externalForce();
 		ThreeDVector* gravity();
 		ThreeDVector* wind();
@@ -59,25 +59,25 @@ class Particle {
 
 		//bool isSurfaceParticle(vector<Particle*>* particles);
 
-		long double color(vector<Particle*>* particles);
-		//long double colorGradient(vector<Particle*>* particles);
+		float color(vector<Particle*>* particles);
+		//float colorGradient(vector<Particle*>* particles);
 
-		static long double colorAt(ThreeDVector* position, vector<Particle*>* particles);
-		static long double colorAt(long double x, long double y, long double z, vector<Particle*>* particles);
+		static float colorAt(ThreeDVector* position, vector<Particle*>* particles);
+		static float colorAt(float x, float y, float z, vector<Particle*>* particles);
 
-		static Particle* createWaterParticle(long double x, long double y, long double z, ThreeDVector* velocity=new ThreeDVector());
-		static Particle* createFogParticle(long double x, long double y, long double z, ThreeDVector* velocity=new ThreeDVector());
-		static Particle* createBoundaryParticle(long double x, long double y, long double z, ThreeDVector* velocity=new ThreeDVector());
+		static Particle* createWaterParticle(float x, float y, float z, ThreeDVector* velocity=new ThreeDVector());
+		static Particle* createFogParticle(float x, float y, float z, ThreeDVector* velocity=new ThreeDVector());
+		static Particle* createBoundaryParticle(float x, float y, float z, ThreeDVector* velocity=new ThreeDVector());
 
 
 		//Poly 6 Kernel from http://www.matthiasmueller.info/publications/sca03.pdf
-		static long double poly6Kernel(long double r, long double h);
+		static float poly6Kernel(float r, float h);
 
 		//Vicosity Kernel from http://www.matthiasmueller.info/publications/sca03.pdf
-		static long double viscosityGradientSquaredKernel(long double r, long double h);
+		static float viscosityGradientSquaredKernel(float r, float h);
 
 		//Spiky Kernel from http://www.matthiasmueller.info/publications/sca03.pdf
-		static ThreeDVector* spikyGradientKernel(ThreeDVector* r, ThreeDVector* r_particle, long double h);
+		static ThreeDVector* spikyGradientKernel(ThreeDVector* r, ThreeDVector* r_particle, float h);
 
 		static void clearColorMap();
